requirements.txt
openai
streamlit
faiss-cpu
PyMuPDF
python-dotenv
tiktoken
